% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepfuncs.R
\name{get_rasters}
\alias{get_rasters}
\title{This is a main function to request fishing effort data from the Global Fishing Watch API
and turn the data into a dataframe containing the information of
vessel name, year, month, day, fishing effort, latitude, longitude, gear type and vessel flag}
\usage{
get_rasters(
  spatial_resolution,
  temporal_resolution,
  group_by = "vessel_id",
  begin,
  end,
  selected_region,
  gear_type = "all",
  region_source = "user_json",
  key
)
}
\arguments{
\item{spatial_resolution}{is required to select either low, equivalent to 0.1 degree,
or high, equivalent to 0.01 degree}

\item{temporal_resolution}{is required to select either daily or monthly or yearly
it is recommended to use daily}

\item{group_by}{is designed to group the data by vessel ID in this version}

\item{begin}{is required to provide the beginning year}

\item{end}{is required to provide the end year}

\item{selected_region}{is required to provide selected fishing regions}

\item{gear_type}{by default, the function selects all gear types but it can be selected
by assigning the gear_type to any of the following strings,
"trawlers", "drifting_longlines", "fixed_gear", "set_longlines", "pots_and_traps"
"set_gillnets", "purse_seines", "other_purse_seines", "tuna_purse_seines","pole_and_line"}

\item{region_source}{is assigned to be "user_json" in this version}

\item{key}{is required to provide the API access token,
visit globalfishingwatch.org/our-apis for more details regarding registration}
}
\value{
Returns a dataframe containing the information of
vessel name, year, month, day, fishing effort, latitude, longitude, gear type and vessel flag
}
\description{
This is a main function to request fishing effort data from the Global Fishing Watch API
and turn the data into a dataframe containing the information of
vessel name, year, month, day, fishing effort, latitude, longitude, gear type and vessel flag
}
\examples{
effort_df = get_rasters(spatial_resolution = "high",
                        temporal_resolution = "daily",
                        begin = 2012, end = 2016,
                        selected_region = c("3L","3N"),
                        key = "Your secret API access token")
# Example returns
head(effort_df)
 vessel_name year month day   effort   lat   long gear_type flag positionStr
    VESSEL1 2012    08  13 0.975833 48.61 -49.86  trawlers  CAN 48.61-49.86
    VESSEL1 2012    08  13 0.608056 48.63 -49.87  trawlers  CAN 48.63-49.87
    VESSEL1 2012    08  16 4.123889 48.61 -49.82  trawlers  CAN 48.61-49.82
    VESSEL1 2012    08  19 1.110833 48.60 -49.78  trawlers  CAN  48.6-49.78
    VESSEL1 2012    08  14 1.472222 48.60 -49.81  trawlers  CAN  48.6-49.81
    VESSEL1 2012    08  17 1.234167 48.59 -49.79  trawlers  CAN 48.59-49.79

}
