% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepfuncs.R
\name{make_annual_ls}
\alias{make_annual_ls}
\title{This function takes the output of the function get_rasters() to create
a list of fishing effort data for different years that is helpful for
data visualization when using the function make_bubmap() in this package}
\usage{
make_annual_ls(data)
}
\arguments{
\item{data}{Takes the output of the function get_rasters() as input}
}
\value{
Returns a list of fishing effort data for different years
}
\description{
This function takes the output of the function get_rasters() to create
a list of fishing effort data for different years that is helpful for
data visualization when using the function make_bubmap() in this package
}
\examples{
effort_df = get_rasters(spatial_resolution = "high",
                        temporal_resolution = "daily",
                        begin = 2012, end = 2016,
                        selected_region = c("3L","3N"),
                        key = "Your secret API access token")

effort_ls = make_annual_ls(effort_df)
# Returns:
effort_ls
[[1]]
    positionStr N    effort   lat   long year
1   39.01-46.55 1  1.705278 39.01 -46.55 2012
2   39.01-47.43 1  1.771389 39.01 -47.43 2012
3   39.01-47.72 1  3.224722 39.01 -47.72 2012
4   39.01-47.77 1  1.270556 39.01 -47.77 2012

[[2]]
    positionStr N    effort   lat   long year
1   39-47.29    1 10.806389 39.00 -47.29 2013
2   39-47.39    1  1.774444 39.00 -47.39 2013
3   39-47.46    1  1.609444 39.00 -47.46 2013
4   39.01-46.55 1  1.656944 39.01 -46.55 2013
...
#'
}
