% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavizfuncs.R
\name{make_bubmap}
\alias{make_bubmap}
\title{This is a main function to create a list of bubble maps for fishing effort
data requested from the Global Fishing Watch API}
\usage{
make_bubmap(
  data,
  selected_map_region,
  begin,
  end,
  alpha = 0.3,
  over_zero_color = "#002E94",
  over_zero_fill = "#002E94",
  zero_color = "#1A1A40",
  zero_fill = "#1A1A40",
  legend_title,
  plot_tilte = "Fishing effort map",
  zeronote = "Weight = 0"
)
}
\arguments{
\item{data}{is required to plug in the output of the function get_rasters()}

\item{selected_map_region}{is required to provide selected regions for the map background.

The current version accepts any of region codes in the FAO Major Fishing Area 21

(i.e., the Northwest Atlantic). The region codes include c("0A", "0B", "1A", "1B", "1C",

"1D", "1E", "1F", "2G", "2H", "2J", "3K", "3L", "3M", "3N", "3O", "3Pn", "3Ps", "4R",

"4S", "4T", "4Vn", "4Vs", "4W", "4X", "5Y", "5Ze", "5Zw", "6A", "6B", "6C", "6D", "6E",

"6F", "6G", "6H."}

\item{begin}{is required to provide the beginning year}

\item{end}{is required to provide the end year}

\item{alpha}{is assigned to be 0.3 by default, and alterable}

\item{over_zero_color}{is assigned to be "#002E94" by default, and alterable,
presenting data points of effort > 0}

\item{over_zero_fill}{is assigned to be "#002E94" by default, and alterable,
presenting data points of effort > 0}

\item{zero_color}{is assigned to be "#1A1A40" by default, and alterable,
presenting data points of effort = 0}

\item{zero_fill}{is assigned to be "#1A1A40" by default, and alterable,
presenting data points of effort = 0}

\item{legend_title}{presents the legend title, taking string inputs}

\item{plot_tilte}{is assigned to be "Fishing effort map" by default,
and alterable, taking string inputs}

\item{zeronote}{is assigned to be "Weight = 0" by default, and alterable,
presenting the legend title for data points of effort = 0}
}
\value{
Returns a list of bubble maps presenting fishing effort for different years
}
\description{
This is a main function to create a list of bubble maps for fishing effort
data requested from the Global Fishing Watch API
}
\examples{
effort_map = make_bubmap(data = effort_ls,
                         selected_map_region = c("3K", "3L", "3N", "3M", "3O", "3Ps", "3Pn"),
                         begin = begin, end = end,
                         legend_title = "Fishing effort\n (hour)",
                         zeronote = "")


}
