% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepfuncs.R
\name{make_region_json}
\alias{make_region_json}
\title{Extracts the coordinates information and turn them into json data
to use in the function get_raster()}
\usage{
make_region_json(selected_region)
}
\arguments{
\item{selected_region}{}
}
\value{
Returns a json file containing coordinates information of the
selected fishing regions that is required in the function get_rasters()
}
\description{
Extracts the coordinates information and turn them into json data
to use in the function get_raster()
}
\examples{
make_region_json(c("21.3.L","21.3.N"))
# Returns:
"{\"geojson\":{\"type\":\"Polygon\",\"coordinates\":[ [[-53.7389,48.9973],
[-53.7425,49.0046],[-53.7475,49.0045],[-53.7484,49.0037],[-53.7559,49.0037],
[-53.7567,49.0029],[-53.7717,49.0029],[-53.7717,49.0037],[-53.7667,49.0046],
[-53.765,49.0071],[-53.7634,49.0062],[-53.76,49.0062],...]]}}"

}
