% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepfuncs.R
\name{make_selected_region}
\alias{make_selected_region}
\title{Creates selected fishing regions}
\usage{
make_selected_region(region)
}
\arguments{
\item{region}{}
}
\value{
Returns the geometry information of the selected fishing regions
}
\description{
Creates selected fishing regions
}
\examples{
make_selected_region(c("3L","3N"))
# Returns:
Simple feature collection with 2 features and 5 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -54.5 ymin: 39 xmax: -46.5 ymax: 49.25
Geodetic CRS:  NAD83
A tibble: 2 × 6
AREA PERIMETER NAFO_ NAFO_ID ZONE                                              geometry
<dbl>     <dbl> <int>   <int> <chr>                                        <POLYGON [°]>
1  23.0      48.9   529       0 3L    ((-53.73892 48.99732, -53.74253 49.00458, -53.74747…
2  31.0      22.4   594       0 3N    ((-50.76667 46, -50.75 46, -50.73333 46, -50.71667 …

}
