% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepfuncs.R
\name{make_year_range}
\alias{make_year_range}
\title{Creates begin and end year objects from the provided date information
that is required in the function get_rasters()}
\usage{
make_year_range(begin, end)
}
\arguments{
\item{begin}{the beginning year}

\item{end}{the end year}
}
\value{
Returns list of strings containing begin and end years
}
\description{
Creates begin and end year objects from the provided date information
that is required in the function get_rasters()
}
\examples{
make_year_range(2012,2018)


}
