% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprepfuncs.R
\name{make_selected_region}
\alias{make_selected_region}
\title{Creates selected fishing regions}
\usage{
make_selected_region(region)
}
\arguments{
\item{region}{}
}
\value{
Returns the geometry information of the selected fishing regions
}
\description{
Creates selected fishing regions
}
\examples{
make_selected_region(c("21.3.L","21.3.N"))
# Returns:
Simple feature collection with 3 features and 15 fields
Geometry type: MULTIPOLYGON
Dimension:     XY
Bounding box:  xmin: -54.5 ymin: 39 xmax: -46.5 ymax: 49.25
Geodetic CRS:  WGS 84
# A tibble: 3 × 16
F_CODE F_LEVEL  F_STATUS OCEAN    SUBOCEAN F_AREA F_SUBA…¹ F_DIV…² F_SUB…³ F_SUB…⁴    ID
<chr>  <chr>       <dbl> <chr>    <chr>    <chr>  <chr>    <chr>   <chr>   <chr>   <int>
1 21.3.O DIVISION        1 Atlantic 2        21     21.3     21.3.O  NA      NA        201
2 21.3.N DIVISION        1 Atlantic 2        21     21.3     21.3.N  NA      NA        218
3 21.3.L DIVISION        1 Atlantic 2        21     21.3     21.3.L  NA      NA        239

}
